Move the docs CRDT tree read path into @wafflebase/docs
Turning a docs Yorkie root back into a `Document` means walking
`doc > block > inline > text`, and for tables `block > row > cell >
block > ...`. That walk lived only in the backend's `docs-tree.ts`,
which a browser cannot import — it is a NestJS module.

The revision-history preview needs exactly the same walk, over a parsed
snapshot instead of a live proxy. Copying it would give the two readers
independent opinions about a format they must agree on, so it moves to
`@wafflebase/docs` and `readDocsRoot` delegates to it. The writer and
`readPageSetup` stay behind: `readPageSetup` has to walk a Yorkie proxy
property by property to dodge its double-encoding.

The shared module is deliberately CRDT-library-agnostic — it reads a
structural `DocsTreeNode` rather than the SDK's `ElementNode` — so
`@wafflebase/docs` gains no `@yorkie-js/sdk` dependency and the two node
dialects can be reconciled by their callers. It is exported from the
node entry too, which is what the backend actually resolves.

No behavior change: `readDocsRoot` returns what it did before.
