+++ packages/docs/src/node.ts
@@ -104,6 +104,12 @@ export {
   serializeMarginFromEdgeAttrs,
   parseMarginFromEdgeAttr,
 } from './model/crdt-attrs.js';
+// The tree→`Document` read path, for the same reason and with the same
+// safety: `model/crdt-tree.js` imports only `model/types.js` and
+// `model/crdt-attrs.js`, and is deliberately CRDT-library-agnostic. The
+// backend's `readDocsRoot` delegates its whole walk to it.
+export { docsTreeToDocument, treeNodeToBlock } from './model/crdt-tree.js';
+export type { DocsTreeNode } from './model/crdt-tree.js';
 export type { StyleId, NamedStyleDef, DocStyles } from './model/named-styles.js';
 export {
   BUILTIN_STYLES,
